Add files via upload
I'm using another project (setting up a data-logging Arduino to examine the  power generation of a solar PV panel) and hope to adapt it to deal with up to 10 amps and from 50-300 volts and further incorporate a photo- detector to log the insolation simultaneously. This data will be used to compare a cooled and non-cooled PV panel.